pyproject.toml
[project]
name = "codepilot"
version = "0.1.0"
description = "Private repository coding agent MVP with CLI, FastAPI, indexing, safe tools, and metrics."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-openai>=1.2.1",
    "langgraph>=1.1.9",
    "mcp>=1.27.0",
    "onnxruntime<1.24",
    "prometheus-client>=0.25.0",
    "pydantic>=2.13.3",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "redis>=7.4.0",
    "rich>=15.0.0",
    "sqlalchemy>=2.0.49",
    "sqlmodel>=0.0.38",
    "typer>=0.24.2",
    "uvicorn>=0.46.0",
]

[project.scripts]
codepilot = "codepilot.cli.main:app"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["codepilot*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
